Remove folder uploads

+++ wp-content/themes/london/framework/extensions/ReduxCore/framework.php
@@ -74,7 +74,7 @@
             // Please update the build number with each push, no matter how small.
             // This will make for easier support when we ask users what version they are using.
 
-            public static $_version = '3.5.2';
+            public static $_version = '3.5.4.6';
             public static $_dir;
             public static $_url;
             public static $_upload_dir;
@@ -103,7 +103,7 @@
 
                         $is_plugin = false;
                         foreach ( get_plugins() as $key => $value ) {
-                            if ( strpos( $key, 'redux-framework.php' ) !== false ) {
+                            if ( is_plugin_active( $key ) && strpos( $key, 'redux-framework.php' ) !== false ) {
                                 self::$_dir = trailingslashit( Redux_Helpers::cleanFilePath( WP_CONTENT_DIR . '/plugins/' . plugin_dir_path( $key ) . 'ReduxCore/' ) );
                                 $is_plugin  = true;
                             }
@@ -170,7 +170,6 @@
             public $toHide = array(); // Values to hide on page load
             public $typography = null; //values to generate google font CSS
             public $import_export = null;
-            public $debug = null;
             public $no_panel = array(); // Fields that are not visible in the panel
             private $show_hints = false;
             public $hidden_perm_fields = array(); //  Hidden fields specified by 'permissions' arg.
@@ -180,6 +179,7 @@
             public $filesystem = null;
             public $font_groups = array();
             public $lang = "";
+            public $dev_mode_forced = false;
 
             /**
              * Class Constructor. Defines the args for the theme options class
@@ -348,9 +348,6 @@
 
                     // Display admin notices in dev_mode
                     if ( true == $this->args['dev_mode'] ) {
-                        require_once( self::$_dir . 'inc/debug.php' );
-                        $this->debug = new ReduxDebugObject ( $this );
-
                         if ( true == $this->args['update_notice'] ) {
                             add_action( 'admin_init', array( $this, '_update_check' ) );
                         }
@@ -403,7 +400,7 @@
 
                     if ( $this->args['dev_mode'] == true || Redux_Helpers::isLocalHost() == true ) {
                         include_once 'core/dashboard.php';
-                        
+
                         if ( ! isset ( $GLOBALS['redux_notice_check'] ) ) {
                             include_once 'core/newsflash.php';
 
@@ -456,7 +453,7 @@
                     // menu title/text
                     'page_title'                => '',
                     // option page title
-                    'page_slug'                 => '_options',
+                    'page_slug'                 => '',
                     'page_permissions'          => 'manage_options',
                     'menu_type'                 => 'menu',
                     // ('menu'|'submenu')
@@ -539,8 +536,9 @@
                         ),
                     ),
                     'show_import_export'        => true,
+                    'show_options_object'       => false,
                     'dev_mode'                  => true,
-                    'system_info'               => false,
+                    
                     'disable_tracking'          => false,
                     'templates_path'            => '',
                     // Path to the templates file for various Redux elements
@@ -567,14 +565,6 @@
                     'meta'   => array( 'class' => 'redux-network-admin' )
                 );
                 $wp_admin_bar->add_node( $args );
-            }
-
-            private function stripslashes_deep( $value ) {
-                $value = is_array( $value ) ?
-                    array_map( 'stripslashes_deep', $value ) :
-                    stripslashes( $value );
-
-                return $value;
             }
 
             public function save_network_page() {
@@ -1122,6 +1112,62 @@
                 return $value;
             }
 
+            public function field_default_values($field) {
+                // Detect what field types are being used
+                if ( ! isset ( $this->fields[ $field['type'] ][ $field['id'] ] ) ) {
+                    $this->fields[ $field['type'] ][ $field['id'] ] = 1;
+                } else {
+                    $this->fields[ $field['type'] ] = array( $field['id'] => 1 );
+                }
+                if ( isset ( $field['default'] ) ) {
+                    $this->options_defaults[ $field['id'] ] = $field['default'];
+                } elseif ( ( $field['type'] != "ace_editor" ) ) {
+                    // Sorter data filter
+
+                    if ( isset( $field['data'] ) && !empty( $field['data'] ) ) {
+                        if (!isset($field['args'])) {
+                            $field['args'] = array();
+                        }
+                        if ( is_array( $field['data'] ) && !empty( $field['data'] ) ) {
+                            foreach ( $field['data'] as $key => $data ) {
+                                if (!empty($data)) {
+                                    if ( ! isset ( $this->field['args'][ $key ] ) ) {
+                                        $field['args'][ $key ] = array();
+                                    }
+                                    $field['options'][ $key ] = $this->get_wordpress_data( $data, $field['args'][ $key ] );
+                                }
+                            }
+                        } else {
+                            $field['options'] = $this->get_wordpress_data( $field['data'], $field['args'] );
+                        }
+                    }
+
+                    if ( $field['type'] == "sorter" && isset ( $field['data'] ) && ! empty ( $field['data'] ) && is_array( $field['data'] ) ) {
+                        if ( ! isset ( $field['args'] ) ) {
+                            $field['args'] = array();
+                        }
+                        foreach ( $field['data'] as $key => $data ) {
+                            if ( ! isset ( $field['args'][ $key ] ) ) {
+                                $field['args'][ $key ] = array();
+                            }
+                            $field['options'][ $key ] = $this->get_wordpress_data( $data, $field['args'][ $key ] );
+                        }
+                    }
+
+                    if (isset ( $field['options'] )) {
+                        if ( $field['type'] == "sortable" ) {
+                            $this->options_defaults[ $field['id'] ] = array();
+                        } elseif ( $field['type'] == "image_select" ) {
+                            $this->options_defaults[ $field['id'] ] = '';
+                        } elseif ( $field['type'] == "select" ) {
+                            $this->options_defaults[ $field['id'] ] = '';
+                        } else {
+                            $this->options_defaults[ $field['id'] ] = $field['options'];
+                        }
+                    }
+                }
+            }
+
             /**
              * Get default options into an array suitable for the settings API
              *
@@ -1158,39 +1204,7 @@
                                     $field['class'] .= "redux-section-indent-start";
                                     $this->sections[ $sk ]['fields'][ $k ] = $field;
                                 }
-                                // Detect what field types are being used
-                                if ( ! isset ( $this->fields[ $field['type'] ][ $field['id'] ] ) ) {
-                                    $this->fields[ $field['type'] ][ $field['id'] ] = 1;
-                                } else {
-                                    $this->fields[ $field['type'] ] = array( $field['id'] => 1 );
-                                }
-                                if ( isset ( $field['default'] ) ) {
-                                    $this->options_defaults[ $field['id'] ] = $field['default'];
-                                } elseif ( isset ( $field['options'] ) && ( $field['type'] != "ace_editor" ) ) {
-                                    // Sorter data filter
-
-                                    if ( $field['type'] == "sorter" && isset ( $field['data'] ) && ! empty ( $field['data'] ) && is_array( $field['data'] ) ) {
-                                        if ( ! isset ( $field['args'] ) ) {
-                                            $field['args'] = array();
-                                        }
-                                        foreach ( $field['data'] as $key => $data ) {
-                                            if ( ! isset ( $field['args'][ $key ] ) ) {
-                                                $field['args'][ $key ] = array();
-                                            }
-                                            $field['options'][ $key ] = $this->get_wordpress_data( $data, $field['args'][ $key ] );
-                                        }
-                                    }
-
-                                    if ( $field['type'] == "sortable" ) {
-                                        $this->options_defaults[ $field['id'] ] = array();
-                                    } elseif ( $field['type'] == "image_select" ) {
-                                        $this->options_defaults[ $field['id'] ] = '';
-                                    } elseif ( $field['type'] == "select" ) {
-                                        $this->options_defaults[ $field['id'] ] = '';
-                                    } else {
-                                        $this->options_defaults[ $field['id'] ] = $field['options'];
-                                    }
-                                }
+                                $this->field_default_values($field);
                             }
                         }
                     }
@@ -1220,6 +1234,33 @@
                 if ( $this->args['global_variable'] == "" && $this->args['global_variable'] !== false ) {
                     $this->args['global_variable'] = str_replace( '-', '_', $this->args['opt_name'] );
                 }
+
+                // Force dev_mode on WP_DEBUG = true and if it's a local server
+                if ( Redux_Helpers::isLocalHost() || ( Redux_Helpers::isWpDebug() ) ) {
+                    if ( $this->args['dev_mode'] != true ) {
+                        $this->args['update_notice'] = false;
+                    }
+                    $this->dev_mode_forced  = true;
+                    $this->args['dev_mode'] = true;
+                }
+                
+                if ($this->args['dev_mode']) {
+                    $this->args['show_options_object'] = true;
+                }
+                
+                // Auto create the page_slug appropriately
+                if ( empty( $this->args['page_slug'] ) ) {
+                    if ( ! empty( $this->args['display_name'] ) ) {
+                        $this->args['page_slug'] = sanitize_html_class( $this->args['display_name'] );
+                    } else if ( ! empty( $this->args['page_title'] ) ) {
+                        $this->args['page_slug'] = sanitize_html_class( $this->args['page_title'] );
+                    } else if ( ! empty( $this->args['menu_title'] ) ) {
+                        $this->args['page_slug'] = sanitize_html_class( $this->args['menu_title'] );
+                    } else {
+                        $this->args['page_slug'] = str_replace( '-', '_', $this->args['opt_name'] );
+                    }
+                }
+
             }
 
             /**
@@ -1361,6 +1402,10 @@
                                     continue;
                                 }
 
+                                if ( isset( $section['permissions'] ) && ! current_user_can( $section['permissions'] ) ) {
+                                    continue;
+                                }
+
                                 add_submenu_page(
                                     $this->args['page_slug'], $section['title'], $section['title'], $this->args['page_permissions'], $this->args['page_slug'] . '&tab=' . $k,
                                     //create_function( '$a', "return null;" )
@@ -1370,16 +1415,6 @@
 
                             // Remove parent submenu item instead of adding null item.
                             remove_submenu_page( $this->args['page_slug'], $this->args['page_slug'] );
-                        }
-
-                        if ( true == $this->args['dev_mode'] ) {
-                            $this->debug->add_submenu();
-                        }
-
-                        if ( true == $this->args['system_info'] ) {
-                            add_submenu_page(
-                                $this->args['page_slug'], __( 'System Info', 'redux-framework' ), __( 'System Info', 'redux-framework' ), $this->args['page_permissions'], $this->args['page_slug'] . '&tab=system_info_default', '__return_null'
-                            );
                         }
                     }
                 }
@@ -1901,7 +1936,18 @@
                     $th = '<div class="redux_field_th">' . $th . '</div>';
                 }
 
-                if ( $this->args['default_show'] === true && isset ( $field['default'] ) && isset ( $this->options[ $field['id'] ] ) && $this->options[ $field['id'] ] != $field['default'] && $field['type'] !== "info" && $field['type'] !== "group" && $field['type'] !== "section" && $field['type'] !== "editor" && $field['type'] !== "ace_editor" ) {
+                $filter_arr = array(
+                    'editor',
+                    'ace_editor',
+                    'info',
+                    'section',
+                    'repeater',
+                    'color_scheme',
+                    'social_profiles',
+                    'css_layout'
+                );
+
+                if ( $this->args['default_show'] == true && isset ( $field['default'] ) && isset ( $this->options[ $field['id'] ] ) && $this->options[ $field['id'] ] != $field['default'] && ! in_array( $field['type'], $filter_arr ) ) {
                     $th .= $this->get_default_output_string( $field );
                 }
 
@@ -1915,7 +1961,7 @@
              * @access      public
              * @return      void
              */
-            public function _register_settings() {
+            public function _register_settings($setOption = true) {
 
                 // TODO - REMOVE
                 // Not used by new sample-config, but in here for legacy builds
@@ -1924,10 +1970,13 @@
                     include( ABSPATH . "wp-includes/pluggable.php" );
                 }
 
-                register_setting( $this->args['opt_name'] . '_group', $this->args['opt_name'], array(
-                    $this,
-                    '_validate_options'
-                ) );
+                //if ($setOption) {
+                    register_setting( $this->args['opt_name'] . '_group', $this->args['opt_name'], array(
+                        $this,
+                        '_validate_options'
+                    ) );
+                //}
+
 
                 if ( is_null( $this->sections ) ) {
                     return;
@@ -2060,8 +2109,10 @@
                                 }
                             }
                             if ( isset ( $field['customizer_only'] ) && $field['customizer_only'] == true ) {
-                                //$display = false;
+                                $display = false;
                             }
+
+
 
                             if ( isset ( $field['permissions'] ) ) {
 
@@ -2102,7 +2153,7 @@
 
                             // Set the defaults to the value if not present
                             $doUpdate = false;
-                            
+
                             // Check fields for values in the default parameter
                             if ( ! isset ( $this->options[ $field['id'] ] ) && isset ( $field['default'] ) ) {
                                 $this->options_defaults[ $field['id'] ] = $this->options[ $field['id'] ] = $field['default'];
@@ -2244,13 +2295,17 @@
                                 if ( isset ( $field['hidden'] ) && $field['hidden'] ) {
                                     $field['label_for'] = 'redux_hide_field';
                                 }
+                                //if ( $setOption ) {
+                                    add_settings_field(
+                                        "{$fieldk}_field", $th, array(
+                                        &$this,
+                                        '_field_input'
+                                    ), "{$this->args['opt_name']}{$k}_section_group", "{$this->args['opt_name']}{$k}_section", $field
+                                    );
+                                //} else {
+                                //    $this->field_head[$field['id']] = $th;
+                                //}
 
-                                add_settings_field(
-                                    "{$fieldk}_field", $th, array(
-                                    &$this,
-                                    '_field_input'
-                                ), "{$this->args['opt_name']}{$k}_section_group", "{$this->args['opt_name']}{$k}_section", $field
-                                );
 
                             }
                         }
@@ -2650,10 +2705,11 @@
                     return;
                 }
 
-                if ( defined( 'WP_CACHE' ) && WP_CACHE && class_exists( 'W3_ObjectCache' ) ) {
+                if ( defined( 'WP_CACHE' ) && WP_CACHE && class_exists( 'W3_ObjectCache' ) && function_exists ( 'w3_instance' ) ) {
                     //echo "here";
-                    $w3  = W3_ObjectCache::instance();
-                    $key = $w3->_get_cache_key( $this->args['opt_name'] . '-transients', 'transient' );
+                    $w3_inst    = w3_instance('W3_ObjectCache');
+                    $w3         = $w3_inst->instance();
+                    $key        = $w3->_get_cache_key( $this->args['opt_name'] . '-transients', 'transient' );
                     //echo $key;
                     $w3->delete( $key, 'transient', true );
                     //set_transient($this->args['opt_name'].'-transients', $this->transients);
@@ -2678,7 +2734,7 @@
 
                 if ( ! empty ( $_POST['data'] ) && ! empty ( $redux->args['opt_name'] ) ) {
 
-                    $values        = array();
+                    $values = array();
                     //if (function_exists('get_magic_quotes_gpc') && get_magic_quotes_gpc()) {
                     //    $process = array(&$_GET, &$_POST, &$_COOKIE, &$_REQUEST);
                     //    while (list($key, $val) = each($process)) {
@@ -2699,7 +2755,7 @@
                     $values = $values[ $redux->args['opt_name'] ];
 
 
-                    if ( function_exists('get_magic_quotes_gpc') && get_magic_quotes_gpc() ) {
+                    if ( function_exists( 'get_magic_quotes_gpc' ) && get_magic_quotes_gpc() ) {
                         $values = array_map( 'stripslashes_deep', $values );
                     }
